pyproject.toml
[tool.poetry]
name = "inferno3-database"
version = "0.1.0"
description = ""
authors = ["Bruno Mouta <brumouta@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
web_server = "scripts:web_server"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pyyaml = "^6.0.1"
pymongo = "^4.4.1"
requests = "^2.31.0"
motor = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
